Adding some clarity to the recurly initializer
Some people tend to be confused by the naming here so I think adding
these comments might be helpful. I also think we should remove the
recurly.js v2 parameter as it seems to be confusing people. If a
developer is generating this file then we should assume they are
starting a new project and thus want them to use recurly.js v3