Remove the site-id constraint from Client
A client no longer needs to provide the site id, it can be inferred from
the API key. It's now optional when initializing the client.

I also added a way to optionally pass it into any operation. This will
set us up for when we have oauth and one authenticated client can
access many sites.