V2 add support for Apply Credit Balance to existing invoice feature
Add new available_credit_balance_in_cents attribute to
AccountBalance. The available_credit_balance_in_cents attribute
is a similar format to the balance_in_cents attribute and contains
the total of open balances for credit invoices in each currency.
This value is useful when trying to determine if the customer's
account has any available credit that can be applied to an
existing collectible invoice on the account.

Add new apply_credit_balance method to Invoice. This action will
be available for collectible charge invoices. If the account has
available credit it will be applied to pay down the balance on the
invoice.